+++ Examples/Import-Excel/ImportMultipleSheetsAsArray.ps1
@@ -1,0 +1,7 @@
+Import-Module $PSScriptRoot\..\..\ImportExcel.psd1 -Force
+
+$xlfile = "$PSScriptRoot\yearlySales.xlsx"
+
+$result = Import-Excel -Path $xlfile -WorksheetName * -Raw
+
+$result | Measure-Object