+++ Examples/Import-Excel/ImportMultipleSheetsAsHashtable.ps1
@@ -1,0 +1,9 @@
+Import-Module $PSScriptRoot\..\..\ImportExcel.psd1 -Force
+
+$xlfile = "$PSScriptRoot\yearlySales.xlsx"
+
+$result = Import-Excel -Path $xlfile -WorksheetName *
+
+foreach ($sheet in $result.Values) {
+    $sheet
+}